# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "ApiResponseLogoutResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/LogoutResponse"
          },
          "success": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "ApiResponseTokenResponse": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/TokenResponse"
          },
          "success": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "LoginRequest": {
        "properties": {
          "email": {
            "example": "user@example.com",
            "format": "email",
            "minLength": 1,
            "type": "string"
          },
          "password": {
            "example": "Password123!",
            "minLength": 1,
            "type": "string"
          }
        },
        "required": ["email", "password"],
        "type": "object"
      },
      "LogoutResponse": {
        "properties": {
          "loggedOut": {
            "type": "boolean"
          }
        },
        "type": "object"
      },
...
      "RefreshTokenRequest": {
        "properties": {
          "refreshToken": {
            "minLength": 1,
            "type": "string"
          }
        },
        "required": ["refreshToken"],
        "type": "object"
      },
...
      "TokenResponse": {
        "properties": {
          "accessToken": {
            "type": "string"
          },
          "refreshToken": {
            "type": "string"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "Bearer Authentication": {
        "bearerFormat": "JWT",
        "name": "Bearer Authentication",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "WealthFlow 투자 플랫폼 API",
    "title": "WealthFlow API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "login",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseTokenResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "로그인",
        "tags": ["Auth"]
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "description": "현재 기기의 refresh token family를 폐기합니다.",
        "operationId": "logout",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseLogoutResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "로그아웃",
        "tags": ["Auth"]
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "description": "RTR 방식으로 refresh token을 회전하고 새 token pair를 발급합니다.",
        "operationId": "refresh",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseTokenResponse"
                }
              }
            },
            "description": "OK"
          }
        },
        "summary": "리프레시 토큰 재발급",
        "tags": ["Auth"]
      }
    },
...
  },
  "servers": [
    {
      "url": "https://d3uib3r331utfe.cloudfront.net"
    }
  ],
...
}